+++ ChangeLog.md
@@ -1,5 +1,19 @@
+## [2.1.0] - 2019-04-03
+### Added
+- Settings panel, colors for rain-addiction (1983097132206)
+- virus quick settings
+- Rain-Addction instance mode (94c2c9305a924)
+- p-settings-color (24a65ed29f)
 
-## [2.0.1] - 2019-04-03
+### Changed
+- when page is not found, play a random iframe (8249e6b340f)
+- sketch at the head of seed.js
+
+### Fixed
+- debouce sketch load rain-addction (94c2c9305a924)
+- play sound before buffer
+
+## [2.0.1] - 2019-04-02
 ### Added
 - Play view : Random button turns into Restart after 5s 
 
@@ -13,7 +27,7 @@
 - Play view JS sketch navigation
 - Kill JS sketch when
 
-## [2.0.0] - 2019-04-03
+## [2.0.0] - 2019-04-02
 ### Added
 - Instance mode : virus, p, eyes (MR: https://github.com/sikaili/skyl.fr/pull/6/)